+++ src/core/Shared/Transaction.cs
@@ -8,18 +8,18 @@ namespace core.Shared
         public Guid EventId { get; }
         public string Ticker { get; }
         public string Description { get; }
-        public double Debit { get; }
-        public double Credit { get; }
+        public decimal Debit { get; }
+        public decimal Credit { get; }
         public DateTimeOffset DateAsDate { get; }
         public string Date => this.DateAsDate.ToString("yyyy-MM-dd");
-        public double Profit => this.Credit - this.Debit;
-        public double ReturnPct
+        public decimal Profit => this.Credit - this.Debit;
+        public decimal ReturnPct
         {
             get
             {
                 if (this.Debit > 0)
                 {
-                    return (this.Credit - this.Debit) * 1.0 / this.Debit;
+                    return (this.Credit - this.Debit) * 1.0m / this.Debit;
                 }
 
                 return 0;
@@ -34,8 +34,8 @@ namespace core.Shared
             Guid eventId,
             string ticker,
             string description,
-            double debit,
-            double credit,
+            decimal debit,
+            decimal credit,
             DateTimeOffset when,
             bool isOption,
             bool isPL)
@@ -51,17 +51,17 @@ namespace core.Shared
             this.IsPL = isPL;
         }
 
-        public static Transaction CreditTx(Guid aggregateId, Guid eventId, string ticker, string description, double credit, DateTimeOffset when, bool isOption)
+        public static Transaction CreditTx(Guid aggregateId, Guid eventId, string ticker, string description, decimal credit, DateTimeOffset when, bool isOption)
         {
             return new Transaction(aggregateId, eventId, ticker, description, 0, credit, when, isOption, false);
         }
 
-        public static Transaction DebitTx(Guid aggregateId, Guid eventId, string ticker, string description, double debit, DateTimeOffset when, bool isOption)
+        public static Transaction DebitTx(Guid aggregateId, Guid eventId, string ticker, string description, decimal debit, DateTimeOffset when, bool isOption)
         {
             return new Transaction(aggregateId, eventId, ticker, description, debit, 0, when, isOption, false);
         }
 
-        public static Transaction PLTx(Guid aggregateId, string ticker, string description, double debit, double credit, DateTimeOffset when, bool isOption)
+        public static Transaction PLTx(Guid aggregateId, string ticker, string description, decimal debit, decimal credit, DateTimeOffset when, bool isOption)
         {
             return new Transaction(aggregateId, Guid.Empty, ticker, description, debit, credit, when, isOption, true);
         }
